Remove query batching
